GridContext gets its own file. Putting it in constants is a bit awkward.
