+++ src/ocean_emulators/constants.py
@@ -1,7 +1,6 @@
-import dataclasses
 import enum
 import logging
-from typing import Self, TypeAlias
+from typing import TypeAlias
 
 logger = logging.getLogger(__name__)
 
@@ -236,31 +235,6 @@ def construct_metadata(data: xr.Dataset) -> dict[str, dict[str, str]]:
 
 class LoaderVersion(enum.Enum):
     OM4_TORCH = "om4-torch"
-
-
-@dataclasses.dataclass(frozen=True)
-class GridContext:
-    """Grid-level context for model forward passes and loss computation.
-
-    Bundles spatial metadata that travels alongside input tensors during training:
-    the ocean/land mask for excluding land cells from gradients, and the grid
-    resolution for latitude-weighted loss calculations.
-
-    Attributes:
-        label_mask: Boolean mask indicating valid ocean cells for each prognostic
-            variable. Shape: (num_prognostic_vars, lat, lon). Land cells are False.
-        input_resolution: Tuple of (latitude, longitude) coordinate tensors defining
-            the spatial grid. Used for cosine-latitude weighting in loss functions.
-    """
-
-    label_mask: PrognosticMask
-    input_resolution: tuple[Lat, Lon]
-
-    def to(self, device: torch.device) -> Self:
-        """Move the label mask to the specified device."""
-        return dataclasses.replace(
-            self, label_mask=self.label_mask.to(device, non_blocking=True)
-        )
 
 
 # TODO(#95): See if this can be removed and replaced.
