+++ src/ocean_emulators/models/fomo.py
@@ -9,11 +9,12 @@ from torch.distributed.algorithms._checkpoint.checkpoint_wrapper import (
     apply_activation_checkpointing,
 )
 
-from ocean_emulators.constants import GridContext, GridSize
+from ocean_emulators.constants import GridSize
 from ocean_emulators.models.base import BaseModel
 from ocean_emulators.models.modules import PerceiverEncoder
 from ocean_emulators.models.modules.encoder import patch_from
 from ocean_emulators.models.modules.unet_backbone import UNetBackbone
+from ocean_emulators.utils.ctx import GridContext
 from ocean_emulators.utils.device import autocast
 
 if TYPE_CHECKING:
